+++ sdk/src/core-ui.ts
@@ -24,6 +24,7 @@ export {
     UINode,
     UIVisual,
     UIVisualType,
+    UIVisualFit,
     UILayoutGeneration,
     UIMask,
     measureText,